backend/requirements.txt
flask
flask-cors
pyjwt
rich
